backend/pyproject.toml
[project]
name = "finance-manager"
version = "0.1.0"
description = "Multi-user budget management system built with Flask and PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "jwt>=1.4.0",
    "psycopg2-binary>=2.9.12",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "werkzeug>=3.1.8",
]

[project.scripts]
finance-manager = "main:main"
